added abstract base class for all non-Jenkins JobManagement implementations
